Update all migrations to avoid loss of data for production mode

+++ database/migrations/orchid/2015_10_19_214425_create_orchid_role_users_table.php
@@ -11,21 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('role_users', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('role_id');
-            $table->primary(['user_id', 'role_id']);
-            $table->foreign('user_id')
-                ->references('id')
-                ->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->foreign('role_id')
-                ->references('id')
-                ->on('roles')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+        if (!Schema::hasTable('role_users')) {
+            Schema::create('role_users', function (Blueprint $table) {
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedInteger('role_id');
+                $table->primary(['user_id', 'role_id']);
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->foreign('role_id')
+                    ->references('id')
+                    ->on('roles')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -33,8 +35,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        if (App::isLocal()) {
-            Schema::dropIfExists('role_users');
-        }
+        // ...
     }
 };
