+++ database/migrations/orchid/2016_08_07_125128_create_orchid_attachmentstable_table.php
@@ -11,38 +11,40 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('attachments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->text('name');
-            $table->text('original_name');
-            $table->string('mime');
-            $table->string('extension')->nullable();
-            $table->bigInteger('size')->default(0);
-            $table->integer('sort')->default(0);
-            $table->text('path');
-            $table->text('description')->nullable();
-            $table->text('alt')->nullable();
-            $table->text('hash')->nullable();
-            $table->string('disk')->default('public');
-            $table->unsignedBigInteger('user_id')->nullable();
-            $table->string('group')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('attachments') && !Schema::hasTable('attachmentable')) {
+            Schema::create('attachments', function (Blueprint $table) {
+                $table->increments('id');
+                $table->text('name');
+                $table->text('original_name');
+                $table->string('mime');
+                $table->string('extension')->nullable();
+                $table->bigInteger('size')->default(0);
+                $table->integer('sort')->default(0);
+                $table->text('path');
+                $table->text('description')->nullable();
+                $table->text('alt')->nullable();
+                $table->text('hash')->nullable();
+                $table->string('disk')->default('public');
+                $table->unsignedBigInteger('user_id')->nullable();
+                $table->string('group')->nullable();
+                $table->timestamps();
+            });
 
-        Schema::create('attachmentable', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('attachmentable_type');
-            $table->unsignedInteger('attachmentable_id');
-            $table->unsignedInteger('attachment_id');
+            Schema::create('attachmentable', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('attachmentable_type');
+                $table->unsignedInteger('attachmentable_id');
+                $table->unsignedInteger('attachment_id');
 
-            $table->index(['attachmentable_type', 'attachmentable_id']);
+                $table->index(['attachmentable_type', 'attachmentable_id']);
 
-            $table->foreign('attachment_id')
-                ->references('id')
-                ->on('attachments')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+                $table->foreign('attachment_id')
+                    ->references('id')
+                    ->on('attachments')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -50,9 +52,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        if (App::isLocal()) {
-            Schema::drop('attachmentable');
-            Schema::drop('attachments');
-        }
+        // ...
     }
 };
